+++ middleware/ratelimiter.go
@@ -1,0 +1,19 @@
+package middleware
+
+import (
+	"time"
+
+	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/fiber/v2/middleware/limiter"
+)
+
+// Set a Rate limiter with a sliding window  - https://docs.gofiber.io/api/middleware/limiter/
+func RateLimiter(app *fiber.App) {
+
+	app.Use(limiter.New(limiter.Config{
+		Max:               20,
+		Expiration:        30 * time.Second,
+		LimiterMiddleware: limiter.SlidingWindow{},
+	}))
+
+}