+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added endpoint healthchecks
 - refactored to a more 'clean' architecture
+- Added a Fiber rate limiter
 
 ## v0.1.0 (2024-01-10)